+++ CHANGELOG.md
@@ -5,6 +5,11 @@ Changes are grouped by sprint weeks, starting each Monday.
 
 ## [Sprint: May 5 - May 9, 2025]
 
+### May 7, 2025
+
+- Add rank progress indicator to account summary
+- change position placement
+
 ### May 6, 2025
 
 - Fixed SuperChain points display and claiming functionality: